Stable Baseline Models + testing

+++ run_agent.py
@@ -114,16 +114,16 @@ if __name__ == '__main__':
     parser.add_argument('--discount', type=float, default=0.99)                         # old default: 0.99
     parser.add_argument('--threshold', type=float, default=20.1)                        # old default: 475
     parser.add_argument('--dropout', type=float, default=0.2)                           # old default: 0.2
-    parser.add_argument('--randomness', action='store_true')                            # old default: False
+    parser.add_argument('--randomness', type=bool, default=False)                       # old default: False
     parser.add_argument('--rendering', action='store_true')
-    parser.add_argument('--overwriting', type=bool, default=True)          # old default: True
+    parser.add_argument('--overwriting', type=bool, default=False)                      # old default: True
     parser.add_argument('--entity', type=str, default='jgu-wandb')
     parser.add_argument('--lr', type=float, default=1e-4)
     parser.add_argument('--test', action='store_true')
-    parser.add_argument('--plot', action='store_true')
+    parser.add_argument('--plot', type=bool, default=True)
     args = parser.parse_args()
 
-    wandb.init(project="box-gym", entity=args.entity)
+    wandb.init(project="box-gym", entity=args.entity, mode="online")                    # mode="online"
     wandb.config = args
 
     if mode == TRAINING:  # train + test new agent
@@ -131,10 +131,13 @@ if __name__ == '__main__':
                                           randomness=args.randomness)  # do_render=True
         input_dim, output_dim = get_env_dims(train_env)
         agent = QAgent(input_dim, output_dim, gamma=args.discount, lr=args.lr)
-        mean_train_rewards, mean_test_rewards = agent.train_loop(train_env, test_env, args, only_testing=args.test)
+        """agent = VanillaGradMLP(input_dim, 100, output_dim, dropout=args.dropout, uses_scale=args.envname=='scale',
+                              scale_exp=args.envname=='scale_exp')"""
+        mean_train_rewards, mean_test_rewards = agent.train_loop(train_env, test_env, args, only_testing=False)
         # save the trained agent
-        with open('agent', 'wb') as agent_file:
-            dill.dump(agent, agent_file)
+        if args.overwriting:   # todo: fix pickling
+            with open('agent', 'wb') as agent_file:
+                dill.dump(agent, agent_file)
         if args.plot:
             plot_rewards(mean_train_rewards, mean_test_rewards, args.threshold)
 
@@ -145,7 +148,7 @@ if __name__ == '__main__':
             # use the loaded agent
             train_env, test_env = create_envs(args.envname, seed=args.seed, do_render=args.rendering,
                                               randomness=args.randomness)
-            _, mean_test_rewards = agent.train_loop(train_env, test_env, args, only_testing=args.test)
+            _, mean_test_rewards = agent.train_loop(train_env, test_env, args, only_testing=True)
         if args.plot:
             plot_test_rewards(mean_test_rewards, args.threshold)
